requirements.txt
# Good Shepherd Python Dependencies
# Updated: November 2025

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
websockets==14.1
slowapi==0.1.9

# Database
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
asyncpg==0.30.0
alembic==1.14.0

# Redis & Caching
redis[hiredis]==5.2.1

# HTTP Client
httpx==0.28.1

# Data Processing
feedparser==6.0.11
python-dateutil==2.9.0

# Environment
python-dotenv==1.0.1

# Search (Meilisearch)
meilisearch==0.31.6

# PDF Generation (optional - install for full PDF support)
# reportlab==4.2.5
# weasyprint==62.3

# Email
aiosmtplib==3.0.2

# Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
psycopg2-binary==2.9.11
aiosqlite==0.21.0

# Observability (OpenTelemetry)
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-logging==0.59b0
opentelemetry-exporter-otlp-proto-grpc==1.38.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx[http2]==0.28.1
